making progress

+++ fundtransferservice/src/main/java/com/example/fundtransferservice/model/entity/TransactionEntity.java
@@ -23,7 +23,7 @@ public class TransactionEntity {
     private TransactionStatus status;
     @Enumerated(EnumType.STRING)
     private TransactionType paymentType;
-    private BigDecimal amount;
+    private double amount;
     private static final BigDecimal plafondMax = BigDecimal.valueOf(1000000);
     private Date issueDate;
     private Date expiryDate;
